requirements.txt
google-adk>=0.1.0
google-genai>=0.3.0
amadeus>=10.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
flask>=3.0.0
flask-cors>=4.0.0